Update and Consolidation
This branch took the original code and re-configured it into a single class, `OSRM`.

It condensed the database accession into a single function that all other functions could pass parameters down to, removing significant amounts of repeated code.
It took individual classes for specific calls and turned them into overloaded functions, allowing for different amounts of options to be passed or handled as defaults. 

It also updated the http requests to current java functions.

This should be the first/last/only commit, as we're abandoning this direction. 
Several extra objects are included for consistency checking - note, none of those checks were implemented.
Additionally, significant functionality was not expanded upon, as it wasn't needed for testing and now won't be needed in the future.
There is an included main function for testing.